Move the interface and document it better
