+++ com.unity.netcode.gameobjects/Runtime/Transports/UTP/INetworkStreamDriverConstructor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cb3c2e99b33d54b4c9507e9550fb1603
